vsx-registry: update api-compatibility handling
The following commit updates the logic present in `@theia/vsx-registry`
which is used to determine API-compatible extensions. The changes update
the previous assumption that vscode-builtin extensions could not be
downloaded through the extensions-view.

The changes improve the performance overall when dealing with `builtin`
extensions by not updating their information through the registry API
calls which are unecessary. As a result, startup times are improved,
local information such as the extension-editor is quicker, and there are
less API-calls overall which do not burden the registry.

The updated logic behaves in the following way:
- builtins bundled with the application are excluded from fetching their
  compatible version.
- builtins installed through the extensions-view are checked for their
  latest compatible version.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>