custom-editor: remove 'editor-preview` coupling
The following commit removes the unnecessary coupling to
`@theia/editor-preview` from `custom-editor` functionality contributed
by the plugin-system.

Adding the coupling was a breaking change for downstream applications
who were not expecting `@theia/editor-preview` from being contributed,
and meant it was difficult to opt-out.

The coupling was introduced to add unused typing so it is safe to
remove.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>