core: implement `fuzzy` searching and highlights
The commit includes updates to both `filterItems` and `findMatches` in
order to make them both `fuzzy` supported.

The method `filterItems` which is used to filter a list of
quick-pick-items based on a filter is now responsible for filtering
items based on `fuzzy` by an item's `label`, `description` or `detail`
(previously only substring `label` matches). The method will now also
properly reset highlights on subsequent filtering, and will apply the
proper highlighting based on which item property it matched against.

The method `findMatches` was also updated in order to support fuzzy, and
it will now properly compute highlights and no longer compute them only
if they are substring matches.

The commit also includes test-cases to verify the behavior of both
methods under different conditions and edge-cases.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>